Fix issue 6
(https://github.com/fangohr/introduction-to-python-for-computational-science-and-engineering/issues/6)

Problem was that the '-' in j-1 was not the normal minus but some
other type of hyphen, which in the translation to LaTeX disappeared.

I also changed the embedded html table to an itemised list; as the
embedded table didn't render in the notebook very well, and neither in
html (only in the latex output).

There is further an issue for markdown (?) parsing the line

 `a[i:j]` returns elements *i* up to *j* − 1

correctly: it makes "up to" italics, not "i" and "j". To avoid that, I
have taken out the italic printing of i and j at this point.