Make pdflatex work
Key was to (i) have an md file (not notebook) as the intro before chapter one,
and this must have a heading that is underlined, i.e.

HEADING
=======

not

Not doing it like this, leads to mixing up of sections (in the notebooks,
written as "# Sectionheading") and chapters in the ouput of pdflatex.